requirements.txt
matplotlib
streamlit
pandas
numpy
matplotlib
seaborn
scikit-learn
pillow